@@ -39,7 +39,7 @@ module.exports = function initContext(
       if (authorizationHeader != null) {
         const basic = authorizationHeader.split(' ');
         if (basic[0].toLowerCase() === 'basic' && basic[1]) {
-          authorizationHeader = new Buffer(basic[1], 'base64').toString('ascii').split(':')[0];
+          authorizationHeader = Buffer.from(basic[1], 'base64').toString('ascii').split(':')[0];
         }
         if (Array.isArray(authorizationHeader)) return authorizationHeader[0];
         return authorizationHeader;        
