@@ -3,25 +3,25 @@
  * Expects settings to be passed the same way as for the main server.
  */
 
-var Cache = require('./Cache'),
-    errorHandling = require('components/errors').errorHandling,
-    utils = require('components/utils');
+const Cache = require('./cache');
+const errorHandling = require('components/errors').errorHandling;
+const utils = require('components/utils');
+const settings = require('./config').load();
+const logger = utils.logging(settings.logs).getLogger('previews-cache-worker');
 
-var settings = require('./config').load(),
-    logger = utils.logging(settings.logs).getLogger('previews-cache-worker');
-var cache = new Cache({
+const cache = new Cache({
   rootPath: settings.eventFiles.previewsDirPath,
-  maxAge: (settings.eventFiles.previewsCacheMaxAge / 1000 || 60 * 60 * 24 * 7) / 1000, // 1w
+  maxAge: (settings.eventFiles.previewsCacheMaxAge / 1000 || 60 * 60 * 24 * 7) / 1000, // 1w
   logger: logger
 });
 
 logger.info('Starting clean-up in ' + settings.eventFiles.previewsDirPath);
-cache.cleanUp(function (err) {
-  if (err) {
+cache.cleanUp()
+  .then(() => {
+    logger.info('Clean-up successful.');
+    process.exit(0);
+  })
+  .catch(err => {
     errorHandling.logError(err, null, logger);
     process.exit(1);
-    return;
-  }
-  logger.info('Clean-up successful.');
-  process.exit(0);
-});
+  });
